Update tests and gsn relayhub script

+++ scripts/test.sh
@@ -7,11 +7,15 @@ set -o errexit
 trap cleanup EXIT
 
 cleanup() {
-  # Kill the ganache-cli instance that we started (if we started one and if it's still running).
+  # Kill the GSN relay server that we started (if we started one and if it's still running).
+  if [ -n "$gsn_relay_server_pid" ] && ps -p $gsn_relay_server_pid > /dev/null; then
+    kill $gsn_relay_server_pid
+  fi
+
+  # Kill the ganache instance that we started (if we started one and if it's still running).
   if [ -n "$ganache_pid" ] && ps -p $ganache_pid > /dev/null; then
     kill -9 $ganache_pid
   fi
-  kill $gsn_relay_server_pid
 }
 
 if [ "$SOLIDITY_COVERAGE" = true ]; then
@@ -19,8 +23,6 @@ if [ "$SOLIDITY_COVERAGE" = true ]; then
 else
   ganache_port=8545
 fi
-
-relayer_port=8099
 
 ganache_running() {
   nc -z localhost "$ganache_port"
@@ -41,6 +43,10 @@ start_ganache() {
     --account="0xd999042bfc9743927b214d2a9be92e32320edffb2457f2c77e51ed1bd6539c09,1000000000000000000000000"
     --account="0xd999042bfc9743927b214d2a9be92e32320edffb2457f2c77e51ed1bd6539c10,1000000000000000000000000"
     --account="0xd999042bfc9743927b214d2a9be92e32320edffb2457f2c77e51ed1bd6539c11,1000000000000000000000000"
+      # 3 accounts to be used for GSN matters.
+    --account="0x956b91cb2344d7863ea89e6945b753ca32f6d74bb97a59e59e04903ded14ad00,1000000000000000000000000"
+    --account="0x956b91cb2344d7863ea89e6945b753ca32f6d74bb97a59e59e04903ded14ad01,1000000000000000000000000"
+    --account="0x956b91cb2344d7863ea89e6945b753ca32f6d74bb97a59e59e04903ded14ad02,1000000000000000000000000"
   )
 
   if [ "$SOLIDITY_COVERAGE" = true ]; then
@@ -55,8 +61,9 @@ start_ganache() {
 setup_gsn_relay() {
   echo "Launching GSN relay server"
 
-  ganache_url="http://localhost:$ganache_port "
-  gsn_relay_server_pid=$(npx oz-gsn run-relayer --ethereumNodeURL $ganache_url --port $relayer_port --detach --quiet)
+  ganache_url="http://localhost:$ganache_port"
+  relayer_port=8099
+  gsn_relay_server_pid=$(npx oz-gsn run-relayer --ethereumNodeURL $ganache_url --port $relayer_port --from "0xbb49ad04422f9fa6a217f3ed82261b942f6981f7" --detach --quiet)
   
   echo "GSN relay server launched!"
 }
